+++ lib/faasm/counter.h
@@ -6,7 +6,7 @@
 namespace faasm {
     void initCounter(FaasmMemory *memory, const char* counterKey);
 
-    uint8_t getCounter(FaasmMemory *memory, const char* counterKey);
+    int getCounter(FaasmMemory *memory, const char* counterKey);
 
     void incrementCounter(FaasmMemory *memory, const char* counterKey) ;
 }
